add readFrom, index creation, and gid for apis (#31)
