+++ .github/workflows/ci-go-lint.yaml
@@ -1,4 +1,4 @@
-name: ci-go-tests
+name: ci-go-lint
 
 on:
   pull_request:
@@ -16,7 +16,7 @@ permissions:
   actions: write
 
 jobs:
-  ci-go-tests:
+  ci-go-lint:
     runs-on: ubuntu-24.04
 
     steps:
