+++ api_test.go
@@ -10,12 +10,14 @@ import (
 )
 
 type User struct {
-	Uid  uint64 `json:"uid,omitempty"`
-	Name string `json:"name,omitempty"`
-	Age  int    `json:"age,omitempty"`
+	Gid     uint64 `json:"gid,omitempty"`
+	Name    string `json:"name,omitempty"`
+	Age     int    `json:"age,omitempty"`
+	ClerkId string `json:"clerk_id,omitempty" db:"constraint=unique"`
 }
 
 func TestCreateApi(t *testing.T) {
+	ctx := context.Background()
 	db, err := modusdb.New(modusdb.NewDefaultConfig(t.TempDir()))
 	require.NoError(t, err)
 	defer db.Close()
@@ -23,41 +25,57 @@ func TestCreateApi(t *testing.T) {
 	db1, err := db.CreateNamespace()
 	require.NoError(t, err)
 
-	require.NoError(t, db1.DropData(context.Background()))
+	require.NoError(t, db1.DropData(ctx))
 
 	user := &User{
-		Name: "B",
-		Age:  20,
+		Name:    "B",
+		Age:     20,
+		ClerkId: "123",
 	}
 
-	uid, _, err := modusdb.Create(context.Background(), db1, user)
+	gid, _, err := modusdb.Create(db, user, db1.ID())
 	require.NoError(t, err)
 
 	require.Equal(t, "B", user.Name)
-	require.Equal(t, uint64(2), uid)
-	require.Equal(t, uint64(2), user.Uid)
+	require.Equal(t, uint64(2), gid)
+	require.Equal(t, uint64(2), user.Gid)
 
 	query := `{
 		me(func: has(User.name)) {
 			uid
 			User.name
 			User.age
+			User.clerk_id
 		}
 	}`
-	resp, err := db1.Query(context.Background(), query)
+	resp, err := db1.Query(ctx, query)
 	require.NoError(t, err)
-	require.JSONEq(t, `{"me":[{"uid":"0x2","User.name":"B","User.age":20}]}`, string(resp.GetJson()))
+	require.JSONEq(t, `{"me":[{"uid":"0x2","User.name":"B","User.age":20,"User.clerk_id":"123"}]}`,
+		string(resp.GetJson()))
 
 	// TODO schema{} should work
-	resp, err = db1.Query(context.Background(), `schema(pred: [User.name, User.age]) {type}`)
+	schemaQuery := `schema(pred: [User.name, User.age, User.clerk_id]) 
+	{
+		type
+		index
+		tokenizer
+	}`
+	resp, err = db1.Query(ctx, schemaQuery)
 	require.NoError(t, err)
 
 	require.JSONEq(t,
-		`{"schema":[{"predicate":"User.age","type":"int"},{"predicate":"User.name","type":"string"}]}`,
+		`{"schema":
+			[
+				{"predicate":"User.age","type":"int"},
+				{"predicate":"User.clerk_id","type":"string","index":true,"tokenizer":["exact"]},
+				{"predicate":"User.name","type":"string"}
+			]
+		}`,
 		string(resp.GetJson()))
 }
 
 func TestCreateApiWithNonStruct(t *testing.T) {
+	ctx := context.Background()
 	db, err := modusdb.New(modusdb.NewDefaultConfig(t.TempDir()))
 	require.NoError(t, err)
 	defer db.Close()
@@ -65,19 +83,20 @@ func TestCreateApiWithNonStruct(t *testing.T) {
 	db1, err := db.CreateNamespace()
 	require.NoError(t, err)
 
-	require.NoError(t, db1.DropData(context.Background()))
+	require.NoError(t, db1.DropData(ctx))
 
 	user := &User{
 		Name: "B",
 		Age:  20,
 	}
 
-	_, _, err = modusdb.Create[*User](context.Background(), db1, &user)
+	_, _, err = modusdb.Create[*User](db, &user, db1.ID())
 	require.Error(t, err)
 	require.Equal(t, "expected struct, got ptr", err.Error())
 }
 
 func TestGetApi(t *testing.T) {
+	ctx := context.Background()
 	db, err := modusdb.New(modusdb.NewDefaultConfig(t.TempDir()))
 	require.NoError(t, err)
 	defer db.Close()
@@ -85,19 +104,21 @@ func TestGetApi(t *testing.T) {
 	db1, err := db.CreateNamespace()
 	require.NoError(t, err)
 
-	require.NoError(t, db1.DropData(context.Background()))
+	require.NoError(t, db1.DropData(ctx))
 
 	user := &User{
 		Name: "B",
 		Age:  20,
 	}
 
-	_, _, err = modusdb.Create(context.Background(), db1, user)
+	gid, _, err := modusdb.Create(db, user, db1.ID())
 	require.NoError(t, err)
 
-	userQuery, err := modusdb.Get[User](context.Background(), db1, uint64(2))
+	gid, queriedUser, err := modusdb.Get[User](db, gid, db1.ID())
 
 	require.NoError(t, err)
-	require.Equal(t, 20, userQuery.Age)
-	require.Equal(t, "B", userQuery.Name)
+	require.Equal(t, uint64(2), gid)
+	require.Equal(t, uint64(2), queriedUser.Gid)
+	require.Equal(t, 20, queriedUser.Age)
+	require.Equal(t, "B", queriedUser.Name)
 }
