+++ api_types.go
@@ -1,0 +1,9 @@
+package modusdb
+
+type UniqueField interface {
+	uint64 | ConstrainedField
+}
+type ConstrainedField struct {
+	Key   string
+	Value any
+}
